combine export/import panel with its properties
